fix args type

+++ src/gym/training_etfstsm.py
@@ -198,8 +198,8 @@ if __name__ == "__main__":
     parser.add_argument('--utility', type=str, help='Utility for the strategy returns evaluation.', default="Sharpe")
     parser.add_argument('--functional', type=str, help='Functional to aggregate across bootstrap samples.', default="means")
     parser.add_argument('--alpha', type=float, help='Confidence level for the rank of the estimates.', default=0.95)
-    parser.add_argument('--k', type=float, help='Number of bootstrap samples.', default=10)
-    parser.add_argument('--cpu_count', type=float, help='Number of CPUs to parallelize process.', default=-1)
+    parser.add_argument('--k', type=int, help='Number of bootstrap samples.', default=10)
+    parser.add_argument('--cpu_count', type=int, help='Number of CPUs to parallelize process.', default=-1)
 
     args = parser.parse_args()
